Add function game
Also added the helper function gameOver to display a message after the
game has finished.

The player is prompted for new input if the input is not one of
Rock, Paper or Scissors.

If the player and computer make the same selection, the round is not
counted and played again.

If the player or the computer reach a score of three, the games is over.
This means that at least 3 rounds are played, but not necessarily more.